Modified site pull_stats for single site
Previous version read in rtp files for a year, accumulated, and then
broke out individual sites. This is better for reading the input files
but blew up memory for the storage arrays. This version works
memory-wise but requires opening each input rtp file 20 times so,
watch for file collisions when running on cluster..